Split buffer submod
